requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# AI / LLM
anthropic==0.39.0

# PDF/Document parsing
PyPDF2==3.0.1
python-docx==1.1.0

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.16.0
httpx==0.25.2
lxml==4.9.3

# Data Processing
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0

# Async Support
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
tenacity==8.2.3
parallel-web
